Bug 699967: Fix crash with wide format output to color PCL.
Writing to wide format output was causing uncompressed lines of
more than 32K to be written to a 32K buffer.

I now recognise that there is an inherent limitation in PCL where
image data can't be larger than 32K, so we'll have to split
page output into subimages and hope they register well enough.

This new commit does that (and solves the overwrite). I am seeing
problems when feeding the output from this into gpcl due to the
delta compression. We believe this is a bug in gpcl, and is being
investigated as bug 699969.